datastore: changes in Core kit datastore
Updated the `commit_transaction` function to enable committing
metadata from pending transactions. In commit transaction we will first
commit metadata and then pending keys to correctly perform the check to
identify if key exists or not.

The strength handling among pending and committed transaction is as:
If pending metadata is strong and committed metadata is weak, commit
the pending setting.

If pending metadata is weak, committed metadata is strong and
the setting is already available, do not downgrade from strong to
weak. Otherwise add the strength file with value as weak.

If pending and committed metadata are the same, no action is
performed.

Additionally, made minor changes to metadata functions for improved
access and flexibility:
- Introduced a `committed` field to dynamically access metadata in
 pending transactions.
- Replaced the hardcoded use of live committed metadata with this
 committed variable ans pass Committed::Live from previous usages.

Refer commit:
https://github.com/bottlerocket-os/bottlerocket-core-kit/pull/294/commits/20a435e3f06625f6c75ccf8109a4a722a055d229
Refer PR:
https://github.com/bottlerocket-os/bottlerocket-core-kit/pull/294